+++ pymetawear/client.py
@@ -86,12 +86,14 @@ class MetaWearClient(object):
     :param str address: A Bluetooth MAC address to a MetaWear board.
     :param str backend: Either ``pygatt`` or ``pybluez``, designating which
         BLE communication backend that should be used.
+    :param float timeout: Timeout for connecting to the MetaWear board. If
+        ``None`` timeout defaults to the backend default.
     :param bool debug: If printout of all sent and received
         data should be done.
 
     """
 
-    def __init__(self, address, backend='pygatt', debug=False):
+    def __init__(self, address, backend='pygatt', timeout=None, debug=False):
         """Constructor."""
         self._address = address
         self._debug = debug
@@ -99,10 +101,10 @@ class MetaWearClient(object):
 
         if backend == 'pygatt':
             self._backend = PyGattBackend(
-                self._address, debug=debug)
+                self._address, timeout=timeout, debug=debug)
         elif backend == 'pybluez':
             self._backend = PyBluezBackend(
-                self._address, debug=debug)
+                self._address, timeout=timeout, debug=debug)
         else:
             raise PyMetaWearException("Unknown backend: {0}".format(backend))
 
@@ -136,8 +138,6 @@ class MetaWearClient(object):
     def __repr__(self):
         return "<MetaWearClient, {0}>".format(self._address)
 
-    # Connection methods
-
     @property
     def backend(self):
         """The requester object for the backend used.
@@ -158,17 +158,6 @@ class MetaWearClient(object):
         libmetawear.mbl_mw_metawearboard_free(self.board)
         self.backend.disconnect()
 
-    def set_logging_state(self, enabled=False):
-        if enabled:
-            libmetawear.mbl_mw_logging_start(self.board)
-        else:
-            libmetawear.mbl_mw_logging_stop(self.board)
-
-    def download_log(self, n_notifies):
-        libmetawear.mbl_mw_logging_download(self.board, n_notifies)
-
-    # Helper methods
-
     def get_handle(self, uuid, notify_handle=False):
         """Get handle for a characteristic UUID.
 
@@ -180,15 +169,12 @@ class MetaWearClient(object):
         """
         return self.backend.get_handle(uuid, notify_handle=notify_handle)
 
-    def _callback_wrapper(self, data):
-        if (data.contents.type_id == DataTypeId.UINT32):
-            data_ptr = cast(data.contents.value, POINTER(c_uint))
-            if data_ptr.contents.value == 1:
-                print("Switch pressed!")
-            elif data_ptr.contents.value == 0:
-                print("Switch released!")
-            else:
-                raise ValueError("Incorrect data returned.")
+    def _set_logging_state(self, enabled=False):
+        if enabled:
+            libmetawear.mbl_mw_logging_start(self.board)
         else:
-            raise RuntimeError('Incorrect data type id: ' + str(data.contents.type_id))
+            libmetawear.mbl_mw_logging_stop(self.board)
+
+    def _download_log(self, n_notifies):
+        libmetawear.mbl_mw_logging_download(self.board, n_notifies)